project2Client
Finished creating the 'second page' where the user can fill out a form
with an issue they are having with their pc. They can input the type of
computer they have, the model number of their computer, the date
they are filling out the form and a text box where they can write a
brief description of the issue they are having. The user should also
be able to view all the tickets they have created, update a ticket and
remove a ticket as well.